fix margin

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/category/presentation/viewholder/AddEditProductCategoryViewHolder.kt
@@ -117,7 +117,7 @@ class AddEditProductCategoryViewHolder(
                 LinearLayout.LayoutParams.WRAP_CONTENT,
                 WEIGHT_ITEM_VIEW
         )
-        params.marginStart = marginStart
+        params.marginStart = marginStart.toInt()
         itemView.tvCategoryNameParent.layoutParams = params
     }
